summary

+++ DCGAN_streetview/utils.py
@@ -36,7 +36,7 @@ def merge_images(images, size):
 
 def merge(images, size):
     h, w = images.shape[1], images.shape[2]
-    img = np.zeros((h * size[0], w * size[1], 3))
+    img = np.zeros((int(h * size[0]), int(w * size[1]), 3))
     for idx, image in enumerate(images):
         i = idx % size[1]
         j = idx // size[1]
